% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_F.R
\name{calc_F}
\alias{calc_F}
\title{Calculate fishing Mortality}
\usage{
calc_F(nSize, nSpecies, mBound, lBound, Ffull, Falpha, FL50, scLinf,
  scLinfMat, phiMin, parameterValues)
}
\arguments{
\item{nSize}{Number of size class intervals species can grow through}

\item{nSpecies}{Number of species in the model}

\item{mBound}{Vector. Mid point of each size class interval}

\item{lBound}{Vector. Lower bound of each size class interval}

\item{Ffull}{Scalar. Fishing mortaliy rate for a fully recruited fish}

\item{Falpha}{Scalar. Number of years to simulate.}

\item{FL50}{Scalar. The length at which 50 \% selection by the fishery occurs.}

\item{scLinf}{Vector. The size class at which each species reaches L_inf (maximum length)}

\item{scLinfMat}{Binary matrix indicating which size classes each species occupies}

\item{phiMin}{Scalar. Model time step (years). See  \code{\link{calc_phi}}}

\item{parameterValues}{Matrix (nSize x nSpecies). See \code{\link{rochet_GB_parameterValues}}}
}
\value{
A matrix is returned

   \item{F}{A matrix (size nSize x nSpecies) of F (fishing mortality) values.
   Note: F = 0 for size classes in which species do not grow (based on von Bertalanfy growth curve)}
}
\description{
Calculates the fishing mortality for each species in each size
class using a logistic selectivity curve.
}
\section{References}{

Hall et al. (2006). A length-based multispecies model for evaluating community responses to fishing. Can. J. Fish. Aquat. Sci. 63:1344-1359.

Rochet et al. (2011). Does selective fishing conserve community biodiversity? Prediction from a length-based multispecies model. Can. J. Fish. Aquat. Sci. 68:469-486
}

