% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_M1.R
\name{calc_M1}
\alias{calc_M1}
\title{Calculates M1 mortality}
\usage{
calc_M1(nSize, nSpecies, lBound, mBound, alphaM1, betaM1, cM1, scLinfMat,
  scLinf, phiMin, parameterValues)
}
\arguments{
\item{nSize}{Number of size class intervals species can grow through}

\item{nSpecies}{Number of species in the model}

\item{lBound}{lower bound of each size class interval}

\item{mBound}{mid point of each size class interval}

\item{alphaM1}{Shape parameter alpha of the beta distribution}

\item{betaM1}{Shape parameter beta of the beta distribution}

\item{cM1}{scales the pdf ??????}

\item{scLinf}{The size class at which each species reaches L_inf (maximum length)}

\item{phiMin}{Model timestep (years). See \code{\link{calc_phi}}}

\item{parameterValues}{data. See \code{\link{data_parameterValues}}}
}
\value{
A matrix is returned

   \code{M1}    - nSize x nSpecies matrix of M1 (natural mortality) values where nSpecies = number of species and nSize = number of size classes.
   Note that M1_i,j = 0 for size classes i in which species j does not reach
}
\description{
calculates other sources of natural mortality other than predation
assumption: follows a beta distribution and that the smaller the size class the greater the mortality
this is based on the ratio of the sizeclass midpoint to the largest sizeclass midpoint (0,1) variable
}
\section{References}{

Hall et al. (2006). A length-based multispecies model for evaluating community responses to fishing. Can. J. Fish. Aquat. Sci. 63:1344-1359.

Rochet et al. (2011). Does selective fishing conserve community biodiversity? Prediction from a length-based multispecies model. Can. J. Fish. Aquat. Sci. 68:469-486
}

