% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_M2_r.R
\name{calc_M2_r}
\alias{calc_M2_r}
\title{calculate predation Mortality}
\usage{
calc_M2_r(nSize, nSpecies, N, ration, suitability, phiMin, otherFood)
}
\arguments{
\item{nSize}{Number of size class intervals species can grow through}

\item{nSpecies}{Number of species in the model}

\item{N}{an nSize x nSpecies matrix of abundance (number of individuals)}

\item{ration}{a list of ration, weight by class, The size class at which each species reaches L_inf (maximum length). see \code{\link{calc_sizepref_suitability}}.}

\item{suitability}{a 3D array of predator prey suitabilities. See \code{\link{calc_ration}}}

\item{phiMin}{Model timestep (years).}

\item{otherfood}{The amount of other food available unaccounted for in the model (grams)}
}
\value{
A matrix is returned

   \code{M2}    - nSize x nSpecies matrix of M2 (predation mortality) values where nSpecies = number of species and nSize = number of size classes.
   Note that M2_i,j = 0 for size classes i in which species j is not preyed upon.
}
\description{
Calculates the predation mortality for each species in each size
class
}
\section{References}{

Hall et al. (2006). A length-based multispecies model for evaluating community responses to fishing. Can. J. Fish. Aquat. Sci. 63:1344-1359.

Rochet et al. (2011). Does selective fishing conserve community biodiversity? Prediction from a length-based multispecies model. Can. J. Fish. Aquat. Sci. 68:469-486
}

\seealso{
\code{\link{calc_sizepref_suitability}}, \code{\link{calc_ration}}
}
