% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_maturity.R
\name{calc_maturity}
\alias{calc_maturity}
\title{Calculates Maturity}
\usage{
calc_maturity(nSize, nSpecies, mBound, scLinfMat, scLinf, parameterValues)
}
\arguments{
\item{nSize}{Number of size class intervals species can grow through}

\item{nSpecies}{Number of species in the model}

\item{mBound}{Mid point of each size class interval}

\item{scLinfMat}{Binary matrix indicating which size classes each species occupies}

\item{scLinf}{The size class at which each species reaches L_inf (maximum length)}

\item{parameterValues}{Data. See \code{\link{rochet_GB_parameterValues}}}
}
\value{
A matrix is returned

   \item{Maturity}{nSize x nSpecies matrix of Maturity proportions where nSpecies = number of species and nSize = number of size classes.
   Note that Maturity_i,j = 0 for size class i in which species j does not reach.}
}
\description{
The proportion of each size class for each species that is mature and contributes to SSB.
}
\section{References}{

Hall et al. (2006). A length-based multispecies model for evaluating community responses to fishing. Can. J. Fish. Aquat. Sci. 63:1344-1359.

Rochet et al. (2011). Does selective fishing conserve community biodiversity? Prediction from a length-based multispecies model. Can. J. Fish. Aquat. Sci. 68:469-486
}

