% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_phi.R
\name{calc_phi}
\alias{calc_phi}
\title{Calculates the proportion of individuals that leave a size class}
\usage{
calc_phi(nSize, nSpecies, uBound, lBound, parameterValues)
}
\arguments{
\item{nSize}{Number of size class intervals species can grow through}

\item{nSpecies}{Number of species in the model}

\item{uBound}{upper bound of each size class interval}

\item{lBound}{lower bound of each size class interval}

\item{parameterValues}{Data. See \code{\link{data_parameterValues}}}
}
\value{
A list is returned

   \code{probGrowOut}    - nSize x nSpecies matrix of probabilities/proportions. Where nSpecies = number of species and nSize = number of size classes.
   Note that probGrowOut_i,j = 0 for size class i in which species j does not reach and probGrowOut_i_j = 0 for largest size class for each species based on Linf.

   \code{phiMin} Model timestep (years). (shortest time it takes any species to grow into nex size class. All species are normalized by this quantity)
}
\description{
The proportion of individuals that leave each size class. Output is scaled relative to the species / size class combinations which is emptied fastest
so that this combination empties completely in one timestep. The time (yrs) for an average fish to grow from the lower to the upper limit of a size class (Hilborn & Walters p428, eqn 13.7.2)
ie scaled to fasted growing individual
}
\section{References}{

Hall et al. (2006). A length-based multispecies model for evaluating community responses to fishing. Can. J. Fish. Aquat. Sci. 63:1344-1359.

Rochet et al. (2011). Does selective fishing conserve community biodiversity? Prediction from a length-based multispecies model. Can. J. Fish. Aquat. Sci. 68:469-486
}

