% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ration.R
\name{calc_ration}
\alias{calc_ration}
\title{Calculate species ration, growth efficiency, and size class weight}
\usage{
calc_ration(nSize, nSpecies, uBound, lBound, midBound, phiMin,
  parameterValues)
}
\arguments{
\item{nSize}{Number of size class intervals species can grow through}

\item{nSpecies}{Number of species in the model}

\item{uBound}{Upper bound of each size class interval}

\item{lBound}{Lower bound of each size class interval}

\item{phiMin}{Model time step (years). See \code{\link{calc_phi}}}

\item{mBound}{Mid point of each size class interval}
}
\value{
A list is returned

   \item{ration}{matrix of size nSize x nSpecies. Amount consumed to account for growth (growth in time interval)/growth efficiency}

   \item{wgt}{matrix of size nSize x nSpecies. Weight of average sized fish in each size class. Uses length weight relationship. See \code{rochet_GB_parameterValues}}

   \item{gEff}{matrix of size nSize x nSpecies. Growth Efficiencies}

   \item{scLinf}{The size class at which each species reaches L_inf (maximum length)}
}
\description{
Calculates the ration, weight, and growth efficiency for a species in a size class.
Ration is defined as the amount consumed to account for growth (growth increment / growth efficiency)
}
\section{References}{

Hall et al. (2006). A length-based multispecies model for evaluating community responses to fishing. Can. J. Fish. Aquat. Sci. 63:1344-1359.

Rochet et al. (2011). Does selective fishing conserve community biodiversity? Prediction from a length-based multispecies model. Can. J. Fish. Aquat. Sci. 68:469-486
}

