% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeMANS.R
\docType{package}
\name{LeMANS}
\alias{LeMANS}
\alias{LeMANS-package}
\title{LeMANS: Length-based Multispecies Analysis by Numerical Simulation}
\description{
An implementation of the LeMANS model as described in the Hall et al (2006) and Rochet et al (2011) publications.
}
\details{
\itemize{
\item Each component of the model (M1, M2, F, suitability, size preference etc) has been modularized
   such that a users code block can be substitued.
\item original data files (used in Rochet et al. (2011)) are bundled with the package.
\item The code is substantially faster (~9 times) than the original MATLAB implementation
\item It serves as a framework for extending the model
}

To learn more about using \code{LeMANS}, start with the vignette: \code{browseVignettes(package="LeMANS")}
}
