% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_foodweb.R
\docType{data}
\name{rochet_GB_foodweb}
\alias{rochet_GB_foodweb}
\title{Foodweb: who eats whom}
\format{A data frame of size 22 x 22
\describe{The list of species can be found in \code{\link{rochet_GB_species}} and in the references below
}}
\source{
Hall et al. (2006). A length-based multispecies model for evaluating community responses to fishing. Can. J. Fish. Aquat. Sci. 63: 1344-1359

Rochet et al. (2011). Does selective fishing conserve community biodiversiy? Predictions from a length-based multispecies model. Can. J. Fish. Aquat. Sci. 68: 469-486
}
\usage{
rochet_GB_foodweb
}
\description{
A 22 species foodweb. Entries are 0 or 1 indicating if species in column eats species in row
}
\seealso{
\code{\link{rochet_GB_foodweb}}, \code{\link{rochet_GB_initialValues}}, \code{\link{rochet_GB_parameterValues}}, \code{\link{rochet_GB_species}}, \code{\link{rochet_GB_modelSetup}}
}
\keyword{datasets}
