% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_M1.R
\name{calc_M1}
\alias{calc_M1}
\title{Calculates M1 mortality}
\usage{
calc_M1(nSize, nSpecies, lBound, mBound, alphaM1, betaM1, cM1, scLinfMat,
  scLinf, phiMin, parameterValues)
}
\arguments{
\item{nSize}{Number of size class intervals species can grow through}

\item{nSpecies}{Number of species in the model}

\item{lBound}{Lower bound of each size class interval}

\item{mBound}{Mid point of each size class interval}

\item{alphaM1}{Shape parameter (alpha) of the beta distribution}

\item{betaM1}{Shape parameter (beta) of the beta distribution}

\item{cM1}{Scales the pdf ??????}

\item{scLinf}{The size class at which each species reaches L_inf (maximum length)}

\item{phiMin}{Model timestep (years). See \code{\link{calc_phi}}}

\item{parameterValues}{Matrix of species specific parameters. See \code{\link{rochet_GB_parameterValues}}}
}
\value{
A matrix is returned

   \item{M1}{A matrix (size nSize x nSpecies) of M1 (natural mortality) values.
   Note: M1 = 0 for size classes in which a species does not reach/grow in to.}
}
\description{
calculates other sources of natural mortality other than predation.
It is assumed that M1 mortality follows a beta distribution and that in general the small and large size classses suffer the greatest mortality.
}
\section{References}{

Hall et al. (2006). A length-based multispecies model for evaluating community responses to fishing. Can. J. Fish. Aquat. Sci. 63:1344-1359.

Rochet et al. (2011). Does selective fishing conserve community biodiversity? Prediction from a length-based multispecies model. Can. J. Fish. Aquat. Sci. 68:469-486
}

