% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_M2_c}
\alias{calc_M2_c}
\title{Calculates predation mortality, M2, in C++}
\usage{
calc_M2_c(nSize, nSpecies, N, scLinf, ration, wgt, suitability, phiMin,
  otherFood)
}
\arguments{
\item{nSize}{Number of size class intervals species can grow through}

\item{nSpecies}{Number of species in the model}

\item{N}{nSize x nSpecies matrix of abundance (number of individuals)}

\item{scLinf}{The size class at which each species reaches L_inf (maximum length)}

\item{ration}{nSize x nSpecies matrix. (growth in time interval)/growth efficiency. See \code{\link{calc_ration}}}

\item{wgt}{Weight of species at the mid point of each size class (Units: grams). See \code{\link{calc_ration}}}

\item{suitability}{3D array of predator size preference of prey size. See \code{\link{calc_sizepref_suitability}}}

\item{phiMin}{Model timestep (years). See \code{\link{calc_phi}}}

\item{otherFood}{Amount of other food available not accounted for in the model (g)}
}
\value{
A matrix is returned

  \code{M2}    - nSize x nSpecies matrix of M2 (predation mortality) values.
   Note that M2_i,j = 0 for size classes i in which species j is not preyed upon.
}
\description{
Calculates the predation mortality for each species in each size class
}
\section{References}{

Hall et al. (2006). A length-based multispecies model for evaluating community responses to fishing. Can. J. Fish. Aquat. Sci. 63:1344-1359.

Rochet et al. (2011). Does selective fishing conserve community biodiversity? Prediction from a length-based multispecies model. Can. J. Fish. Aquat. Sci. 68:469-486
alc_
}

\seealso{
\code{\link{calc_sizepref_suitability}}, \code{\link{calc_ration}} \code{\link{calc_phi}}
}
