% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_maturity.R
\name{calc_maturity}
\alias{calc_maturity}
\title{Calculates Maturity}
\usage{
calc_maturity(nSize, nSpecies, mBound, scLinfMat, scLinf, parameterValues)
}
\arguments{
\item{nSize}{Number of size class intervals species can grow through}

\item{nSpecies}{Number of species in the model}

\item{mBound}{Mid point of each size class interval}

\item{scLinfMat}{A matrix (nSize x nSpecies) of binary values indicating which size classes each species occupies.}

\item{scLinf}{A vector (length nSpecies) of size classes indicating the maximum size class a species can occupy. Based on L_inf (maximum length)}

\item{parameterValues}{A Matrix of species specific parameters. See \code{\link{rochet_GB_parameterValues}}}
}
\value{
A matrix is returned

   \item{Maturity}{A matrix (nSize x nSpecies) of Maturity proportions.
   Note: Maturity = 0 for size classes in which a species does not reach.}
}
\description{
The proportion of each species (in each size class) that are considered mature.
Matrue indisviduals contribute to SSB.
}
\section{References}{

Hall et al. (2006). A length-based multispecies model for evaluating community responses to fishing. Can. J. Fish. Aquat. Sci. 63:1344-1359.

Rochet et al. (2011). Does selective fishing conserve community biodiversity? Prediction from a length-based multispecies model. Can. J. Fish. Aquat. Sci. 68:469-486
}

