% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_population_growth.R
\name{calc_population_growth}
\alias{calc_population_growth}
\title{Calculates the change in the poulation in each time step}
\usage{
calc_population_growth(nSize, nSpecies, N, probGrowOut)
}
\arguments{
\item{nSize}{Number of size class intervals species can grow through}

\item{nSpecies}{Number of species in the model}

\item{N}{A matrix (nSize x nSpecies) of species abundance (number of individuals)}

\item{probGrowOut}{A matrix (nSize x nSpecies) of proportions.
Note: probGrowOut = 0 for size classes in which a species does not reach and probGrowOut = 0 for largest size class. It is not possible to grow out of the largest size class since it is determined by Linf. See \code{\link{calc_phi}}}
}
\value{
A list is returned

   \item{N}{A matrix (nSize x nSpecies) of abundance (number of individuals)}
}
\description{
Updates the species' population size. This change is a function of the number of individuals that grow out of their present size class and the number that remain.
}
\section{References}{

Hall et al. (2006). A length-based multispecies model for evaluating community responses to fishing. Can. J. Fish. Aquat. Sci. 63:1344-1359.

Rochet et al. (2011). Does selective fishing conserve community biodiversity? Prediction from a length-based multispecies model. Can. J. Fish. Aquat. Sci. 68:469-486
}

