% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ration.R
\name{calc_ration}
\alias{calc_ration}
\title{calculate species ration, growth efficiency, and size class weight}
\usage{
calc_ration(nSize, nSpecies, uBound, lBound, midBound, phiMin,
  parameterValues)
}
\arguments{
\item{nSize}{Number of size class intervals species can grow through}

\item{nSpecies}{Number of species in the model}

\item{uBound}{upper bound of each size class interval}

\item{lBound}{lower bound of each size class interval}

\item{phiMin}{Model timestep (years)}

\item{mBound}{mid point of each size class interval}
}
\value{
A list is returned

   \code{ration}
   \code{wgt}
   \code{gEff}
   \code{scLinf}
}
\description{
Calculates the ration, weight, and growth efficiency for a species in a size class
}
\section{References}{

Hall et al. (2006). A length-based multispecies model for evaluating community responses to fishing. Can. J. Fish. Aquat. Sci. 63:1344-1359.

Rochet et al. (2011). Does selective fishing conserve community biodiversity? Prediction from a length-based multispecies model. Can. J. Fish. Aquat. Sci. 68:469-486
}

