% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sizepref_suitability.R
\name{calc_sizepref_suitability}
\alias{calc_sizepref_suitability}
\title{Calculates size preference and suitabilities}
\usage{
calc_sizepref_suitability(nSize, nSpecies, mBound, spMu, spSigma, wgt,
  scLinf, FW)
}
\arguments{
\item{nSize}{Number of size class intervals species can grow through}

\item{nSpecies}{Number of species in the model}

\item{mBound}{Mid point of each size class interval}

\item{spMu}{Mean of log normal distribution. See \code{\link{rochet_GB_modelSetup}}}

\item{spSigma}{Standard deviation of the log normal distribution. See \code{\link{rochet_GB_modelSetup}}}

\item{wgt}{Weight of species at the mid point of each size class (Units: grams). See \code{\link{calc_ration}}}

\item{scLinf}{The size class at which each species reaches L_inf (maximum length). See \code{\link{calc_ration}}}

\item{FW}{Food web represented as a binary matrix. Predator (columns) and prey (rows). See \code{\link{rochet_GB_foodweb}}}
}
\value{
A list is returned

   \item{sizePref}{3D array of prey size preference for predator size.}

   \item{suitability}{3D array of prey size suitability for predator size. Suitability is calculated as (sizePref*Food Web). These
   values are then standardized across all prey species/size classes for a given predator in a size class}
}
\description{
Size preferences for prey species by predator species are modelld with a lognormal distribution.
The bundled data use a parameterization that gives a maximal preference to prey that are approximately 1/33rd the weight of the predator.
}
\section{Notes on 3D arrays}{


The 3D arrays contain prey size class in the rows, prey species in the columns (nSizeClass x nSpecies). The 1st dimension represents a predator/size class comnination.
For example slice [1,,] represents the preference of predator species 1 in size class 1 to all prey species /size class combinations.
The 1st dimension has length equal to nSpecies*nsizeClass. However many entries are zero determined by each species Linf
}

\section{References}{

Hall et al. (2006). A length-based multispecies model for evaluating community responses to fishing. Can. J. Fish. Aquat. Sci. 63:1344-1359.

Rochet et al. (2011). Does selective fishing conserve community biodiversity? Prediction from a length-based multispecies model. Can. J. Fish. Aquat. Sci. 68:469-486
}

