% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_key_run.R
\name{plot_key_run}
\alias{plot_key_run}
\title{Plots output from the key_run script}
\usage{
plot_key_run(dataSet, title, is.aggregated = T)
}
\arguments{
\item{dataSet}{The data set to be plotted}

\item{title}{The title of the figure}
}
\description{
Plots output from the key_run script
}
